test: add tests for tcs3400 offset functions
Add tests for both get/set offset for the clear and RGB tcs3400 as well
as verifying that it's possible to read the RGB sensor with a null
buffer (since it should never write to the buffer anyway).

BRANCH=none
BUG=none
TEST=twister

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3981204
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Simon Glass <sjg@chromium.org>

+++ util/ectool.cc
@@ -5901,7 +5901,7 @@ static int cmd_motionsense(int argc, char **argv)
 		       resp->perform_calib.offset[0],
 		       resp->perform_calib.offset[1],
 		       resp->perform_calib.offset[2]);
-		if ((uint16_t)resp->perform_calib.temp ==
+		if (resp->perform_calib.temp ==
 		    EC_MOTION_SENSE_INVALID_CALIB_TEMP)
 			printf("Temperature at calibration unknown\n");
 		else
@@ -5963,7 +5963,7 @@ static int cmd_motionsense(int argc, char **argv)
 		       resp->sensor_offset.offset[0],
 		       resp->sensor_offset.offset[1],
 		       resp->sensor_offset.offset[2]);
-		if ((uint16_t)resp->sensor_offset.temp ==
+		if (resp->sensor_offset.temp ==
 		    EC_MOTION_SENSE_INVALID_CALIB_TEMP)
 			printf("temperature at calibration unknown\n");
 		else
